# src/immich_memories/web/openapi.json
{
  "components": {
    "schemas": {
...
      "CutBrief": {
        "description": "What to cut; None leaves a flag out, so the CLI's own default applies.",
        "properties": {
          "all_trips": {
            "default": false,
            "title": "All Trips",
            "type": "boolean"
          },
          "birthday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthday"
          },
          "day": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day"
          },
          "duration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "end": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "exclude_asset": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Exclude Asset",
            "type": "array"
          },
          "from_album": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Album"
          },
          "hemisphere": {
            "anyOf": [
              {
                "enum": [
                  "north",
                  "south"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hemisphere"
          },
          "holiday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holiday"
          },
          "include_asset": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Include Asset",
            "type": "array"
          },
          "include_live_photos": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Live Photos"
          },
          "include_photos": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Photos"
          },
          "memory_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Type"
          },
          "month": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Month"
          },
          "near_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Near Date"
          },
          "people_expression": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "People Expression"
          },
          "period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          },
          "person": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Person",
            "type": "array"
          },
          "person_match": {
            "anyOf": [
              {
                "enum": [
                  "and",
                  "or"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Match"
          },
          "season": {
            "anyOf": [
              {
                "enum": [
                  "spring",
                  "summer",
                  "fall",
                  "autumn",
                  "winter"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "sharing": {
            "anyOf": [
              {
                "enum": [
                  "just-us",
                  "family",
                  "shareable"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharing"
          },
          "start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "trip_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Index"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "years_back": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Back"
          }
        },
        "title": "CutBrief",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobView": {
        "properties": {
          "argv": {
            "items": {
              "type": "string"
            },
            "title": "Argv",
            "type": "array"
          },
          "cancel_requested": {
            "default": false,
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "command": {
            "title": "Command",
            "type": "string"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "cut",
              "render",
              "scan"
            ],
            "title": "Kind",
            "type": "string"
          },
          "meta": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ]
            },
            "default": {},
            "title": "Meta",
            "type": "object"
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "progress": {
            "$ref": "#/components/schemas/JobProgress"
          },
          "result_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Run Id"
          },
          "started_at": {
            "title": "Started At",
            "type": "number"
          },
          "status": {
            "default": "running",
            "enum": [
              "running",
              "succeeded",
              "failed",
              "cancelled",
              "interrupted"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "argv",
          "started_at",
          "command",
          "progress"
        ],
        "title": "JobView",
        "type": "object"
      },
...
      "Recut": {
        "description": "The pool's ticks: pictures the next cut must keep, and pictures it must leave out.",
        "properties": {
          "exclude": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Exclude",
            "type": "array"
          },
          "include": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Include",
            "type": "array"
          }
        },
        "title": "Recut",
        "type": "object"
      },
...
      "ShownCommand": {
        "properties": {
          "command": {
            "title": "Command",
            "type": "string"
          }
        },
        "required": [
          "command"
        ],
        "title": "ShownCommand",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Immich Memories",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cuts": {
      "post": {
        "description": "Cut this brief with `generate --no-render`; the job ends with the run the cut became.",
        "operationId": "start_cut_api_v1_cuts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CutBrief"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Cut",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/cuts/command": {
      "post": {
        "description": "The command this brief stands for, as the page offers it to copy.",
        "operationId": "cut_command_api_v1_cuts_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CutBrief"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShownCommand"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cut Command",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "description": "Where the job is, from its record and the files its child writes.",
        "operationId": "read_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "description": "Stop the job's child, from this tab or any other.",
        "operationId": "cancel_job_api_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/events": {
      "get": {
        "description": "Server-sent events: the job's view on every change, until it finishes.",
        "operationId": "job_events_api_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Events",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/output": {
      "get": {
        "description": "What the child printed, secrets removed: the reason a failed job gives.",
        "operationId": "job_output_api_v1_jobs__job_id__output_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Output",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/roster/scan": {
      "post": {
        "description": "Read the library's people again with `people scan`, keeping every answer given.",
        "operationId": "scan_people_api_v1_roster_scan_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Scan People",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/film": {
      "get": {
        "description": "The rendered film, by byte range so the player can seek.",
        "operationId": "film_api_v1_runs__run_id__film_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Film",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/recut": {
      "post": {
        "description": "Cut the same brief again with the owner's ticks, as `generate --include/--exclude` does.",
        "operationId": "recut_api_v1_runs__run_id__recut_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Recut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recut",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/renders": {
      "post": {
        "description": "Render this run's cut, or one revision of it, with `runs render`.",
        "operationId": "start_render_api_v1_runs__run_id__renders_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderOptions"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Render",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}